full-stack-deepagents/mcp-server-oa/requirements.txt
mcp>=1.26.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.34.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
